backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pymupdf==1.23.7
langchain==0.0.335
langchain-openai==0.0.2
openai>=1.6.1
supabase==1.0.3
prisma==0.11.0
psycopg2-binary==2.9.9
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
tiktoken>=0.5.2